requirements.txt
altgraph==0.17.4
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
setuptools==76.0.0
ttkthemes==3.2.2
